add user login functionality

+++ frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,20 @@
 import { useWorkoutContext } from "../hooks/useWorkoutsContext"
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import { useAuthContext } from "../hooks/useAuthContext"
 
 const WorkoutDetails = ({workout}) => {
     const {dispatch} = useWorkoutContext()
-
+    const {user} = useAuthContext
     const handleClick = async () => {
+        if(!user){
+            return
+        }
+
         const responce = await fetch('/api/workouts/' + workout._id,{
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
 
         })
         const json = await responce.json()
